delivery with qr feature added
