+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "school-services-consthistcert",
-  "version": "2.4.0",
+  "version": "2.5.0",
   "description": "Espacio: SchoolServices CHC microservice",
   "keywords": [
     "loopback-application",
