+++ public/index.html
@@ -84,7 +84,7 @@
   <body>
     <div class="info">
       <h1>school-services-consthistcert</h1>
-      <p>Version 2.4.0</p>
+      <p>Version 2.5.0</p>
 
       <h3>OpenAPI spec: <a href="/openapi.json">/openapi.json</a></h3>
       <h3>API Explorer: <a href="/explorer">/explorer</a></h3>
