+++ src/controllers/index.ts
@@ -2,16 +2,9 @@ export * from './academic-history.controller';
 export * from './academic-record.controller';
 export * from './admission-certificate.controller';
 export * from './calendars.controller';
-export * from './copy-of-academic-program.controller';
-export * from './document-loan.controller';
-export * from './duplicated-tae.controller';
-export * from './good-conduct.controller';
-export * from './document-copy.controller';
 export * from './program-change.controller';
 export * from './proof-of-study.controller';
-export * from './proof-of-study.uteg.controller';
 export * from './report-card.controller';
 export * from './scholarship.controller';
 export * from './social-service.controller';
 export * from './study-certificate.controller';
-export * from './study-certificate.uteg.controller';
